// src/views/certificationReviewCenter/cbam/ReportForm/type.ts
export interface ConfigCNResp {
  /**
   * cn编码
   */
  defaultCode?: string;
  /**
   * cn名称
   */
  defaultName?: string;
  /**
   * 0;1
   */
  delFlag?: number;
  id?: number;
  /**
   * 产品分类id
   */
  productCategoryId?: number;
  [property: string]: any;
}
// ...
export interface EleCalculator {
  cbamId?: number;
  /**
   * 电力排放系数
   */
  coefficient?: number | null;
  /**
   * 0;1
   */
  delFlag?: number;
  /**
   * 电力使用值
   */
  eleValue?: number | null;
  id?: number;
  /**
   * 产品工序id
   */
  productProcessId?: number;
  [property: string]: any;
}
// ...
export interface ProductAttribution {
  /**
   * 报表
   */
  cbamId?: number;
  /**
   * 0;1
   */
  delFlag?: number;
  /**
   * 0无;1单一;2多个
   */
  eleChoose?: number;
  /**
   * 电力来源数据字典
   */
  eleSource?: string;
  /**
   * 元素枚举:1直接排放;2热力的输入输出;3尾气;4电力使用;5电力输出(1:直接排放; 2:热力的输入输出; 3:尾气/隐含排放; 4:电力使用; 5:电力输出/电力排放系数;
   * 10:NON)
   */
  emissionElement?: number;
  /**
   * 来源枚举(1:测量值; 2:默认值; 3:未知)
   */
  emissionSource?: number;
  id?: number;
  /**
   * 输入/使用排放系数;前驱的电力使用
   */
  inputFactor?: number;
  /**
   * 热力/尾气/电力输入;前驱的间接排放
   */
  inputPower?: number;
  /**
   * 是否存在:0无;1是;2不存在(0:无; 1:是; 2:否)
   */
  isExists?: number;
  /**
   * 是否生产工序关联:0是;1否
   */
  isProcess?: number;
  /**
   * 输出/回收排放系数;电力排放系数
   */
  outFactor?: number;
  /**
   * 热力/尾气/电力输出 直接排放量;前驱的直接排放
   */
  outPower?: number;
  /**
   * 生产工序id
   */
  productProcessId?: number;
  /**
   * 更新人
   */
  updateBy?: number;
  /**
   * 更新时间
   */
  updateTime?: Date;
  [property: string]: any;
}